refactor: Layout Root, Header, Body 분리
