+++ src/components/Layout/LayoutBody.tsx
@@ -5,6 +5,7 @@ export const LayoutBody = ({ children }: PropsWithChildren) => (
   <main
     className={css({
       display: 'flex',
+      flexDirection: 'column',
       width: '100%',
       maxWidth: '1200px',
       margin: '60px auto 0 auto',
